Clear up the U+FDF2 confusion
U+FDF2, like most of the Arabic Presentation Forms blocks, is a legacy
compatibility character and should not be used for text input. Turning
the (<alef>)<lam><lam><heh> sequence into name of God ligature is the
job of the font and should be handled by smart font features.

The U+FDF2 handling options and macros are kept to not break existing
documents, but they just do nothing now.